Keypad Change
Changed the keypad, added ICO file. Copied Ian's account file's information in to account.cs (need to talk about what we're planning with this). Error catching added for account if and more than 6 numbers are entered. 0 number on the keypad is not part of the 2d array that was created for the rest of the buttons due to aligning issues. Clear and cancel buttons are working as intended. Menu strip added to main menu.